+++ abc/xyz.lua
@@ -1,0 +1,7 @@
+local Graphene, meta = ...
+
+print("abc.xyz loader:", meta.FSID)
+
+return {
+	Name = "xyz"
+}